refactor(organisms): merge SfTab
