--- packages/vue/src/components/organisms/SfTabs/_internal/SfTab.html
@@ -1,17 +1,0 @@
-<Fragment>
-  <div
-    class="sf-tabs__title"
-    @click="tabClick"
-    :class="{'sf-tabs__title--active': isActive}"
-  >
-    {{title}}
-    <div class="sf-tabs__chevron">
-      <SfChevron :class="{'sf-chevron--top': isActive}" />
-    </div>
-  </div>
-
-  <div class="sf-tabs__content">
-    <!--@slot Default. Here you should pass your tab content-->
-    <slot></slot>
-  </div>
-</Fragment>
