+++ packages/vue/src/components/organisms/SfTabs/_internal/SfTab.vue
@@ -1,5 +1,60 @@
-<template lang="html" src="./SfTab.html"> </template>
-<script src="./SfTab.js"></script>
+<template>
+  <Fragment>
+    <div
+      class="sf-tabs__title"
+      :class="{ 'sf-tabs__title--active': isActive }"
+      @click="tabClick"
+    >
+      {{ title }}
+      <div class="sf-tabs__chevron">
+        <SfChevron :class="{ 'sf-chevron--top': isActive }" />
+      </div>
+    </div>
+
+    <div class="sf-tabs__content">
+      <!--@slot Default. Here you should pass your tab content-->
+      <slot></slot>
+    </div>
+  </Fragment>
+</template>
+<script>
+import { Fragment } from "vue-fragment";
+
+import SfChevron from "../../../atoms/SfChevron/SfChevron.vue";
+
+export default {
+  name: "SfTab",
+  components: {
+    Fragment,
+    SfChevron
+  },
+  props: {
+    /**
+     * Tab title.
+     */
+    title: {
+      type: String,
+      default: "MenuItem"
+    }
+  },
+  data() {
+    return {
+      isActive: false,
+      desktopMin: 1024
+    };
+  },
+  methods: {
+    tabClick() {
+      const width = Math.max(
+        document.documentElement.clientWidth,
+        window.innerWidth
+      );
+      if (this.isActive && width > this.desktopMin) return;
+      this.$parent.$emit("toggle", this._uid);
+    }
+  }
+};
+</script>
 <style>
 /* @import '../../../../utilities/transitions/transitions.scss'; */
 .slide-left-enter-active {
